Fix content handler route pattern
